+++ src/main/java/univ/kgu/carely/domain/team/repository/comment/CommentRepository.java
@@ -1,4 +1,4 @@
-package univ.kgu.carely.domain.team.repository;
+package univ.kgu.carely.domain.team.repository.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import univ.kgu.carely.domain.team.entity.Comment;
