+++ src/main/java/univ/kgu/carely/domain/team/repository/team/CustomTeamRepository.java
@@ -1,4 +1,4 @@
-package univ.kgu.carely.domain.team.repository;
+package univ.kgu.carely.domain.team.repository.team;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
