+++ src/main/java/univ/kgu/carely/domain/team/repository/team/TeamRepository.java
@@ -1,4 +1,4 @@
-package univ.kgu.carely.domain.team.repository;
+package univ.kgu.carely.domain.team.repository.team;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import univ.kgu.carely.domain.team.entity.Team;
